pyproject.toml
[project]
name = "Medical_Search_Pro_Agent"
version = "0.1.0"
description = "Ahsan Ali Gill  ahs462agk@gmail.com"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anyio>=4.9.0",
    "google-adk>=1.1.1",
    "secure-smtplib>=0.1.1",
    "python-dotenv>=1.0.0",
    "serpapi>=0.1.5",
    "requests>=2.32.3",
]
